requirements.txt
python-dotenv==1.0.1
pyTelegramBotAPI==4.22.0
mysql-connector-python==9.0.0
